+++ src/main/java/com/telepathicgrunt/the_bumblezone/modCompat/ProductiveBeesRedirection.java
@@ -8,11 +8,11 @@ import net.minecraftforge.event.entity.living.LivingSpawnEvent;
 import java.util.Random;
 
 /**
- * This class is used so java wont load BeesourcefulCompat class and crash if
+ * This class is used so java wont load ProductiveBeesCompat class and crash if
  * the mod isn't on. This is because java will load classes if their method is
  * present even though it isn't called when going through some code. However,
  * java won't load classes referenced in the method ahead of time so the
- * redirection works to keep BeesourcefulCompat unloaded by "nesting" the method
+ * redirection works to keep ProductiveBeesCompat unloaded by "nesting" the method
  * dependent on the mod.
  */
 public class ProductiveBeesRedirection {